a buncha things to back up

+++ src/main/java/me/madmagic/chemcraft/util/MouseUtil.java
@@ -10,7 +10,7 @@ public class MouseUtil {
         return isMouseOver(mouseX, mouseY, x, y, size, size);
     }
 
-    public static boolean isMouseOver(double mouseX, double mouseY, int x, int y, int width, int height) {
+    public static boolean isMouseOver(double mouseX, double mouseY, float x, float y, float width, float height) {
         return (mouseX >= x && mouseX <= x + width) && (mouseY >= y && mouseY <= y + height);
     }
 }
